Threaded commit
In this commit the game was divided into two threads running in parallel. The game thread starts the game and runs the game loop and all the logic. The tester thread only provides the inputs to the game thread, which when in test mode come from a text file.
Known problems:
- Threads aren't synchronized properly, for example, when running the test command to switch into test mode that command only takes effect on the next input. This happens because the test thread is effectively ahead of the game thread, already trying to process the input from the user before it can know that it should switch into test mode
- Again a synchronization issue, when switching in and out of test mode commands will be executed out of order(?) and will need the user to press enter more than once